Moves types and api for customers in global api
