+++ src/api/customer/index.ts
@@ -1,3 +1,2 @@
-export * from './endpoints';
 export * from './queries';
 export * from './types';
